requirements.txt
gitdb==4.0.4
GitPython==3.1.1